requirements.txt
flask==3.1.0
python-dotenv==1.0.1
pymongo==4.11.1
Werkzeug~=3.1.3
Flask-JWT-Extended~=4.7.1
Flask-Bcrypt==1.0.1
Flask-APScheduler==1.13.1
Flask-Mail==0.10.0
Flask-Pydantic==0.12.0
Flask-PyMongo==3.0.1
stripe===11.5.0